pyproject.toml
[project]
name = "jusk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.1",
    "django-browser-reload>=1.18.0",
    "django-stubs-ext>=5.2.0",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
]

[tool.ruff]
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
]
line-length = 88
indent-width = 4
fix = true

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "PL", # pylint
]
ignore = [
    "E501", # line too long
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
ignore_missing_imports = true
follow_untyped_imports = true
disallow_untyped_calls = false
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
    "tests",
]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.0",
    "djangorestframework-stubs[compatible-mypy]>=3.16.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
]
